feat: DB 5개 테이블 설계 및 사용자 자동 로그인 구현
- DB: users, questions, user_responses, evaluations, study_stats 테이블 생성
- backend: User entity/repository/service/controller 추가 (GET /api/user/current)
- frontend: 앱 시작 시 김학생 자동 로그인, 네비게이션에 사용자 이름 표시
- 샘플 데이터 SQL 추가 (김학생, 목표점수 140)

+++ backend/src/main/java/com/aihye/toeicspeaking/controller/RecordController.java
@@ -24,7 +24,7 @@ public class RecordController {
     }
 
     @PostMapping
-    public ResponseEntity<Map<String, Long>> create(@RequestBody RecordRequest request) {
+    public ResponseEntity<Map<String, Integer>> create(@RequestBody RecordRequest request) {
         PracticeRecord record = recordService.createRecord(request);
         return ResponseEntity.status(HttpStatus.CREATED)
                 .body(Map.of("id", record.getId()));
