+++ backend/src/main/java/com/aihye/toeicspeaking/repository/RecordRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RecordRepository extends JpaRepository<PracticeRecord, Long> {
+public interface RecordRepository extends JpaRepository<PracticeRecord, Integer> {
 
     List<PracticeRecord> findAllByOrderByCreatedAtDesc();
 }
